% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_tidyverse_data.R
\name{upsert_tidyverse_data}
\alias{upsert_tidyverse_data}
\title{upsert_tidyverse_data}
\usage{
upsert_tidyverse_data(con)
}
\arguments{
\item{con}{A connection to postgres}
}
\description{
Takes recent tidyverse download data and stores it in Postgres
}
